# openapi.yaml
swagger: '2.0'
info:
  title: CRM Gateway Config
  description: API Gateway mapping for CRM API
  version: 1.0.0
schemes:
  - https
paths:
...
  /api/paymenttypes/{id}:
    get:
      operationId: get_id_paymenttypes
      parameters: [{ name: id, in: path, required: true, type: string }]
      x-google-backend:
        address: https://crm-api-768989935757.us-central1.run.app/api/paymenttypes/{id}
      responses: { '200': { description: OK } }
    put:
      operationId: put_id_paymenttypes
      parameters: [{ name: id, in: path, required: true, type: string }]
      x-google-backend:
        address: https://crm-api-768989935757.us-central1.run.app/api/paymenttypes/{id}
      responses: { '200': { description: OK } }
    delete:
      operationId: del_id_paymenttypes
      parameters: [{ name: id, in: path, required: true, type: string }]
      x-google-backend:
        address: https://crm-api-768989935757.us-central1.run.app/api/paymenttypes/{id}
      responses: { '200': { description: OK } }
...